+++ dp3t-sdk/sdk/src/main/java/org/dpppt/android/sdk/backend/SignatureVerificationInterceptor.java
@@ -1,3 +1,9 @@
+/*
+ * Created by Ubique Innovation AG
+ * https://www.ubique.ch
+ * Copyright (c) 2020. All rights reserved.
+ */
+
 package org.dpppt.android.sdk.backend;
 
 import androidx.annotation.NonNull;
@@ -10,7 +16,6 @@ import java.util.Arrays;
 
 import org.dpppt.android.sdk.util.SignatureUtil;
 
-import io.jsonwebtoken.security.SignatureException;
 import okhttp3.Interceptor;
 import okhttp3.Response;